game

+++ src/main/java/com/example/application/data/service/GridService.java
@@ -3,6 +3,7 @@ package com.example.application.data.service;
 import com.example.application.data.entity.Grid;
 
 import com.example.application.data.repository.GridRepository;
+import com.example.application.data.repository.LevelRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +14,9 @@ public class GridService {
 
     @Autowired
     GridRepository gridRepository;
+
+    @Autowired
+    LevelRepository levelRepository;
 
     public List<Grid> getAllGrids() {
         return gridRepository.findAll();
@@ -27,7 +31,8 @@ public class GridService {
     }
 
     public boolean deleteGrid(Integer gridId) {
-        if(gridRepository.findById(gridId).isPresent()) {
+        if(gridRepository.findById(gridId).isPresent()
+                && !levelRepository.existsLevelByGrid(gridRepository.findById(gridId).get())) {
             gridRepository.deleteById(gridId);
             return true;
         }
